+++ playground/playground.js
@@ -1,12 +1,13 @@
 import { Loxer, NamedError } from '../dist/index.js';
 
-Loxer.init({ dev: true });
+Loxer.init({ dev: true, defaultLevels: { devLevel: 'info' } });
 Loxer.highlight().log('it works!', '...very well');
 const lox = Loxer.m().open('look how it starts');
 Loxer.of(lox).add('it is beautiful');
 Loxer.of(lox).error('even the errors');
 Loxer.of(lox).close('until the end');
 
+Loxer.warn('this is a warning');
 Loxer.error('this is a string error');
 Loxer.error(404);
 Loxer.error(false);
@@ -18,6 +19,7 @@ Loxer.highlight().error('this is a highlighted error that prints the stack!!!');
 
 const id = Loxer.m().open('This is the opening log');
 Loxer.of(id).add('this is a single added log');
+Loxer.of(id).warn('this is a single added warning');
 Loxer.of(id).error('this is an added error');
 Loxer.of(id).close('this is the closing log');
 
@@ -28,7 +30,3 @@ const extended = new NamedError(
   'this is my custom Error that extends another one',
   existing
 );
-
-console.log('A', [existing.name, existing.message, existing.stack]);
-console.log('A', [custom.name, custom.message, custom.stack]);
-console.log('A', [extended.name, extended.message, extended.stack]);